# fix (performance): #3606 Add cache for pocket stories and topics