This a file on disk as a persistent cache.

Fixes #3606