@Mardak 2017-10-05
Some initial testing, here's number of frames (30fps) from first paint to 1) search box 2) strings 3) topics/stories (+relative times for 2&3 from 1):
```
before: 10, 14, 18/18 ( +4,  +8/ +8)
before: 14, 21, 25/25 ( +7, +11/+11)
before: 13, 17, 20/23 ( +4,  +7/+10)
 after: 12, 17, 17/17 ( +5,  +5/ +5)
 after: 14, 19, 19/19 ( +5,  +5/ +5)
 after: 13, 23, 23/23 (+10, +10/+10)
```
If we just take the median times, it looks like:
```
before: 13, 17, 21/23 ( +4,  +8/+10)
 after: 13, 18, 18/18 ( +5,  +5/ +5)
```
So at least on my machine, with this fix, topics and stories show up at the same time as strings with caching being ~100ms/167ms faster than network. Unclear if the slightly slower strings with this caching is just noise or related.